Add TimeElapsed function and integrate timing in PF_StoreBuffer and PF_Processor for performance measurement Printing send, sor, and wait time in PF_Processor

+++ sources/parallel.h
@@ -251,7 +251,7 @@ static inline size_t sizeof_datatype(MPI_Datatype type)
 /* parallel.c */
 extern LONG PF_allocateSbuf(void);
 extern int PF_allocatePFTerm(int numtasks);
-extern int    PF_ForwardTermsToMaster(void);
+extern LONG    PF_ForwardTermsToMaster(void);
 extern void   PF_SetupFlatRequestsView(void);
 extern int 	  PF_ReducerInit(void);
 extern int    PF_EndSort(void);
